Rename hopscotch_sc_map to bhopscotch_map and hopscotch_sc_set to bhopscotch_set.
